+++ web/src/features/blobstorage-integration/blobstorage-integration-router.ts
@@ -9,7 +9,6 @@ import {
 import { encrypt } from "@langfuse/shared/encryption";
 import { blobStorageIntegrationFormSchema } from "@/src/features/blobstorage-integration/types";
 import { TRPCError } from "@trpc/server";
-import { throwIfNoEntitlement } from "@/src/features/entitlements/server/hasEntitlement";
 import {
   logger,
   BlobStorageIntegrationProcessingQueue,
@@ -27,11 +26,6 @@ export const blobStorageIntegrationRouter = createTRPCRouter({
   get: protectedProjectProcedure
     .input(z.object({ projectId: z.string() }))
     .query(async ({ input, ctx }) => {
-      throwIfNoEntitlement({
-        entitlement: "integration-blobstorage",
-        sessionUser: ctx.session.user,
-        projectId: input.projectId,
-      });
       throwIfNoProjectAccess({
         session: ctx.session,
         projectId: input.projectId,
@@ -65,11 +59,6 @@ export const blobStorageIntegrationRouter = createTRPCRouter({
     .input(blobStorageIntegrationFormSchema.extend({ projectId: z.string() }))
     .mutation(async ({ input, ctx }) => {
       try {
-        throwIfNoEntitlement({
-          entitlement: "integration-blobstorage",
-          sessionUser: ctx.session.user,
-          projectId: input.projectId,
-        });
         throwIfNoProjectAccess({
           session: ctx.session,
           projectId: input.projectId,
@@ -165,11 +154,6 @@ export const blobStorageIntegrationRouter = createTRPCRouter({
     .input(z.object({ projectId: z.string() }))
     .mutation(async ({ input, ctx }) => {
       try {
-        throwIfNoEntitlement({
-          entitlement: "integration-blobstorage",
-          sessionUser: ctx.session.user,
-          projectId: input.projectId,
-        });
         throwIfNoProjectAccess({
           session: ctx.session,
           projectId: input.projectId,
@@ -200,11 +184,6 @@ export const blobStorageIntegrationRouter = createTRPCRouter({
     .input(z.object({ projectId: z.string() }))
     .mutation(async ({ input, ctx }) => {
       try {
-        throwIfNoEntitlement({
-          entitlement: "integration-blobstorage",
-          sessionUser: ctx.session.user,
-          projectId: input.projectId,
-        });
         throwIfNoProjectAccess({
           session: ctx.session,
           projectId: input.projectId,
@@ -282,11 +261,6 @@ export const blobStorageIntegrationRouter = createTRPCRouter({
     .input(z.object({ projectId: z.string() }))
     .mutation(async ({ input, ctx }) => {
       try {
-        throwIfNoEntitlement({
-          entitlement: "integration-blobstorage",
-          sessionUser: ctx.session.user,
-          projectId: input.projectId,
-        });
         throwIfNoProjectAccess({
           session: ctx.session,
           projectId: input.projectId,
