+++ web/src/features/playground/page/hooks/usePlaygroundCache.ts
@@ -1,13 +1,12 @@
 import { useEffect, useState } from "react";
 
 import { type PlaygroundCache } from "../types";
-import { useHasEntitlement } from "@/src/features/entitlements/hooks";
 
 const playgroundCacheKey = "playgroundCache";
 
 export default function usePlaygroundCache() {
   const [cache, setCache] = useState<PlaygroundCache>(null);
-  const available = useHasEntitlement("playground");
+
   const setPlaygroundCache = (cache: PlaygroundCache) => {
     sessionStorage.setItem(playgroundCacheKey, JSON.stringify(cache));
   };
@@ -24,7 +23,7 @@ export default function usePlaygroundCache() {
   }, []);
 
   return {
-    playgroundCache: available ? cache : null,
-    setPlaygroundCache: available ? setPlaygroundCache : () => {},
+    playgroundCache:  cache
+    setPlaygroundCache:  setPlaygroundCache : () => {}
   };
 }
