+++ web/src/features/posthog-integration/posthog-integration-router.ts
@@ -9,17 +9,11 @@ import {
 import { decrypt, encrypt } from "@langfuse/shared/encryption";
 import { posthogIntegrationFormSchema } from "@/src/features/posthog-integration/types";
 import { TRPCError } from "@trpc/server";
-import { throwIfNoEntitlement } from "@/src/features/entitlements/server/hasEntitlement";
 
 export const posthogIntegrationRouter = createTRPCRouter({
   get: protectedProjectProcedure
     .input(z.object({ projectId: z.string() }))
     .query(async ({ input, ctx }) => {
-      throwIfNoEntitlement({
-        entitlement: "integration-posthog",
-        sessionUser: ctx.session.user,
-        projectId: input.projectId,
-      });
       throwIfNoProjectAccess({
         session: ctx.session,
         projectId: input.projectId,
@@ -54,11 +48,6 @@ export const posthogIntegrationRouter = createTRPCRouter({
     .input(posthogIntegrationFormSchema.extend({ projectId: z.string() }))
     .mutation(async ({ input, ctx }) => {
       try {
-        throwIfNoEntitlement({
-          entitlement: "integration-posthog",
-          sessionUser: ctx.session.user,
-          projectId: input.projectId,
-        });
         throwIfNoProjectAccess({
           session: ctx.session,
           projectId: input.projectId,
@@ -101,11 +90,6 @@ export const posthogIntegrationRouter = createTRPCRouter({
     .input(z.object({ projectId: z.string() }))
     .mutation(async ({ input, ctx }) => {
       try {
-        throwIfNoEntitlement({
-          entitlement: "integration-posthog",
-          sessionUser: ctx.session.user,
-          projectId: input.projectId,
-        });
         throwIfNoProjectAccess({
           session: ctx.session,
           projectId: input.projectId,
