+++ web/src/features/public-api/components/LLMApiKeyList.tsx
@@ -26,7 +26,6 @@ import { api } from "@/src/utils/api";
 import { DialogDescription } from "@radix-ui/react-dialog";
 import { Alert, AlertDescription, AlertTitle } from "@/src/components/ui/alert";
 import { CreateLLMApiKeyDialog } from "./CreateLLMApiKeyDialog";
-import { useEntitlements } from "@/src/features/entitlements/hooks";
 
 export function LlmApiKeyList(props: { projectId: string }) {
   const hasAccess = useHasProjectAccess({
@@ -34,26 +33,18 @@ export function LlmApiKeyList(props: { projectId: string }) {
     scope: "llmApiKeys:read",
   });
 
-  // only show if the user has access to features that require LLM API keys
-  const entitlements = useEntitlements();
-  const isAvailable =
-    entitlements.includes("playground") ||
-    entitlements.includes("model-based-evaluations");
-
   const apiKeys = api.llmApiKey.all.useQuery(
     {
       projectId: props.projectId,
     },
     {
-      enabled: hasAccess && isAvailable,
+      enabled: hasAccess,
     },
   );
 
   const hasExtraHeaderKeys = apiKeys.data?.data.some(
     (key) => key.extraHeaderKeys.length > 0,
   );
-
-  if (!isAvailable) return null;
 
   if (!hasAccess) {
     return (
