+++ web/src/pages/project/[projectId]/settings/index.tsx
@@ -15,10 +15,7 @@ import { PostHogLogo } from "@/src/components/PosthogLogo";
 import { Card } from "@/src/components/ui/card";
 import { ScoreConfigSettings } from "@/src/features/scores/components/ScoreConfigSettings";
 import { TransferProjectButton } from "@/src/features/projects/components/TransferProjectButton";
-import {
-  useEntitlements,
-  useHasEntitlement,
-} from "@/src/features/entitlements/hooks";
+import { useHasEntitlement } from "@/src/features/entitlements/hooks";
 import { useHasProjectAccess } from "@/src/features/rbac/utils/checkProjectAccess";
 import { useRouter } from "next/router";
 import { SettingsDangerZone } from "@/src/components/SettingsDangerZone";
@@ -46,11 +43,6 @@ export function useProjectSettingsPages(): ProjectSettingsPage[] {
     "prompt-protected-labels",
   );
 
-  const entitlements = useEntitlements();
-  const showLLMConnectionsSettings =
-    entitlements.includes("playground") ||
-    entitlements.includes("model-based-evaluations");
-
   if (!project || !organization || !router.query.projectId) {
     return [];
   }
@@ -60,7 +52,7 @@ export function useProjectSettingsPages(): ProjectSettingsPage[] {
     organization,
     showBillingSettings,
     showRetentionSettings,
-    showLLMConnectionsSettings,
+    showLLMConnectionsSettings: true,
     showProtectedLabelsSettings,
   });
 }
@@ -250,10 +242,6 @@ export default function SettingsPage() {
 }
 
 const Integrations = (props: { projectId: string }) => {
-  const hasPosthogEntitlement = useHasEntitlement("integration-posthog");
-  const hasBlobStorageEntitlement = useHasEntitlement(
-    "integration-blobstorage",
-  );
   const hasAccess = useHasProjectAccess({
     projectId: props.projectId,
     scope: "integrations:CRUD",
@@ -274,7 +262,6 @@ const Integrations = (props: { projectId: string }) => {
             <ActionButton
               variant="secondary"
               hasAccess={hasAccess}
-              hasEntitlement={hasPosthogEntitlement}
               href={`/project/${props.projectId}/settings/integrations/posthog`}
             >
               Configure
@@ -301,7 +288,6 @@ const Integrations = (props: { projectId: string }) => {
             <ActionButton
               variant="secondary"
               hasAccess={hasAccess}
-              hasEntitlement={hasBlobStorageEntitlement}
               href={`/project/${props.projectId}/settings/integrations/blobstorage`}
             >
               Configure
